requirements.txt
accelerate==0.21.0
addict==2.4.0
bitsandbytes==0.41.0
bleach==6.1.0
einops==0.6.1
einops-exts==0.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.1.3
fastapi==0.100.1
ffmpy==0.3.1
gradio==3.35.2
gradio_client==0.2.9
huggingface-hub==0.16.4
ninja==1.11.1.1
numpy==1.24.2
openai==0.27.8
opencv-python==4.8.0.74
pandas==2.0.3
peft==0.4.0
Pillow==9.4.0
pycocotools==2.0.6
pydantic==1.10.12
ray==2.6.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.2
sentencepiece==0.1.99
shortuuid==1.0.11
spacy==3.7.2
tensorboard==2.14.0
tensorboard-data-server==0.7.1
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
torch==2.1.1
torchmetrics==0.2.0
torchvision==0.16.1
tqdm==4.65.2
transformers==4.31.0
uvicorn==0.23.2
wandb==0.15.11
yarl==1.9.2
deepspeed==0.9.5
